refactor project structure & code formattation
