+++ .prettierrc.json
@@ -1,0 +1,10 @@
+{
+	"trailingComma": "none",
+	"tabWidth": 2,
+	"semi": false,
+	"singleQuote": true,
+	"printWidth": 100,
+	"useTabs": true,
+	"bracketSameLine": true,
+	"quoteProps": "consistent"
+}
