--- src/Loading/Loading.jsx
@@ -1,15 +1,0 @@
-import { useEffect } from "react"
-
-const Loading = (p) => {
-  useEffect(() => {
-    const data = localStorage.getItem('user')
-    if (data) p.setView('menu')
-    else p.setView('login')
-  }, [])
-
-  return (
-    <div></div>
-  )
-}
-
-export default Loading
