--- src/Login/Login.style.css
@@ -1,11 +1,0 @@
-#columnLayout {
-    display: flex;
-    flex-direction: column;
-    padding: 0px 10px;
-}
-
-.menuButton {
-    font-size: 1.2rem;
-    margin-bottom: 2.2rem;
-    padding: 0.8rem;
-}
