+++ src/Pages/ListaNozze/Interactions/index.jsx
@@ -1,0 +1,5 @@
+const Interactions = (p) => {
+	return <div>Interactions</div>
+}
+
+export default Interactions
