+++ src/Pages/Menu/Menu.style.css
@@ -1,0 +1,34 @@
+#columnLayout {
+	display: flex;
+	flex-direction: column;
+	padding: 0px 10px;
+}
+
+.menuButton {
+	font-size: 1.2rem;
+	margin-bottom: 2.2rem;
+	padding: 0.8rem;
+}
+
+#menu_main_container {
+	min-height: 88vh;
+	max-width: 500px;
+	width: 90vw;
+}
+
+.ggCover {
+	width: 100%;
+	border-radius: 1rem;
+}
+
+.menu_body_container {
+	padding: 1rem;
+}
+
+.menu_logout {
+	text-decoration: underline;
+	cursor: pointer;
+	text-align: center;
+	color: var(--colorShadeA);
+	margin-top: 1rem;
+}
