+++ src/UI_kit/Card/Card.style.css
@@ -1,82 +1,90 @@
 .cardImage {
-    height: 100%;
-    cursor: pointer;
+	height: 100%;
+	cursor: pointer;
 }
 
 .cardImageContainer {
-    border: 2px solid var(--colorShadeA);
-    border-radius: 1em;
-    order: 0;
-    width: 12rem;
-    height: 8rem; 
-    overflow: hidden;
-    display: flex;
-    justify-content: center;
+	border: 2px solid var(--colorShadeA);
+	border-radius: 1em;
+	order: 0;
+	width: 12rem;
+	height: 8rem;
+	overflow: hidden;
+	display: flex;
+	justify-content: center;
 }
 
 .cardImageLarge {
-    width: 100%;
-    height: auto;
+	width: 100%;
+	height: auto;
 }
 
 .cardImageLargeContainer {
-    border-radius: 1em;
-    order: 0;
-    width: 100%;
-    overflow: hidden;
-    display: flex;
-    justify-content: center;
-    align-items: center;
-    max-height: auto;
-    height: 20rem;
+	border-radius: 1em;
+	order: 0;
+	width: 100%;
+	overflow: hidden;
+	display: flex;
+	justify-content: center;
+	align-items: center;
+	max-height: auto;
+	height: 20rem;
 }
 
 .cardContainer {
-    display: flex;
-    padding-bottom: 1.5rem;
-    align-items: flex-start;
-    flex-wrap: wrap;
-    justify-content: space-between;
-    margin: 1rem 0rem;
+	display: flex;
+	padding-bottom: 1.5rem;
+	align-items: center;
+	flex-wrap: wrap;
+	justify-content: space-between;
+	margin: 1rem 0rem;
 }
 
 .cardOpenModalButton {
-    order: 2;
+	order: 2;
+	display: flex;
+	width: 5.5rem;
+	justify-content: space-between;
 }
 
 .cardContent {
-    width: 100%;
-    padding: 1rem 1rem;
-    order: 1;
-    max-width: 30rem;
+	width: 100%;
+	padding: 1rem 1rem;
+	order: 1;
+	max-width: 30rem;
 }
 
 .cardTitle {
-    font-weight: bold;
-    font-family: 'Raleway';
-    font-size: 1.1rem;
-    color: var(--colorShadeA);
+	font-weight: bold;
+	font-family: 'Raleway';
+	font-size: 1.1rem;
+	color: var(--colorShadeA);
+}
+
+.card_iconbox {
+	width: 2rem;
+	margin-top: 1rem;
 }
 
 @media only screen and (max-width: 784px) {
-    .cardContent {
-      order: 2;
-      padding: 1rem 0rem;
-      max-width: 33rem;
-    }
-    .cardOpenModalButton {
-        order: 1;
-    }
-  }
+	.cardContent {
+		order: 2;
+		padding: 1rem 0rem;
+		max-width: 33rem;
+	}
+	.cardOpenModalButton {
+		order: 1;
+	}
+}
 
-  @media only screen and (max-width: 700px) {
-    .cardImageLarge {
-        width: auto;
-        height: 20rem;
-    }
-    
-    .cardImageLargeContainer {
-        max-height: 20rem;
-        height: auto;
-    }
-  }
+@media only screen and (max-width: 700px) {
+	.cardImageLarge {
+		width: auto;
+		height: 20rem;
+	}
+
+	.cardImageLargeContainer {
+		max-height: 20rem;
+		height: auto;
+	}
+}
