+++ src/UI_kit/CopyTextCell/index.jsx
@@ -1,4 +1,4 @@
-import React, { useEffect, useState} from 'react';
+import React, { useState} from 'react';
 import { genRandomStr } from '../../utils'
 import './CopyTextCell.css';
