+++ src/UI_kit/GMapsButton/index.jsx
@@ -3,7 +3,7 @@ import './GMapsButton.css';
 const GMapsButton = ({text, address}) => {
   return (
     <div className="GMapsButton_container">
-      <a href={address} target="_blank">
+      <a href={address} target="_blank" rel="noreferrer">
         <div className="GMapsButton_copyButton">
           <i className={`fa fa-map GMapsButton_copyIcon`}></i>
         </div>
