+++ src/UI_kit/Icons/index.jsx
@@ -1,0 +1,17 @@
+export const CommentIcon = () => (
+	<svg version="1.1" x="0px" y="0px" viewBox="0 0 1000 1000" fill="rgb(106, 163, 137)">
+		<g>
+			<path d="M500,806.3c-35.8,0-70.3-4.2-103.5-11.2l-144.2,86.7l1.9-142.1c-110.5-66.5-183-176.3-183-300.9c0-203,192-367.5,428.8-367.5c236.8,0,428.8,164.5,428.8,367.5C928.8,641.7,736.8,806.3,500,806.3L500,806.3z M500,10C229.4,10,10,202,10,438.8c0,135.3,71.8,255.8,183.8,334.4V990l214.7-130.2c29.7,4.9,60.3,7.7,91.6,7.7c270.6,0,490-192,490-428.8C990,202,770.6,10,500,10L500,10z" />
+		</g>
+	</svg>
+)
+
+export const HeartIcon = () => (
+	<svg version="1.1" x="0px" y="0px" viewBox="0 0 1000 1000" fill="rgb(106, 163, 137)">
+		<g>
+			<g transform="translate(0.000000,511.000000) scale(0.100000,-0.100000)">
+				<path d="M2465,4470c-1249-191.8-2174-1123.3-2342.4-2357.3C22.4,1383.7,254.7,682.5,845.1-65.7c72.5-93.8,1038-1076.4,2144.2-2184.7l2012.1-2014.2l2031.2,2035.5C9172.5-84.9,9238.6-14.5,9511.4,450.1c385.8,654.3,483.8,1325.7,296.3,2027c-196.1,731.1-733.2,1389.7-1413.1,1728.6c-194,98-537.1,208.9-771.6,249.4c-306.9,53.3-809.9,32-1108.3-44.8c-541.4-140.7-1046.5-447.6-1374.8-835.5c-70.3-83.1-132.2-151.3-138.5-153.5c-6.4,0-57.5,57.5-115.1,130c-187.6,228.1-535,500.9-839.8,654.3c-187.6,95.9-532.9,206.7-758.8,247.2C3115.1,4482.8,2612,4493.4,2465,4470z M3253.6,3813.5c392.2-83.1,754.5-277.1,1031.6-552c72.5-70.3,260-298.4,415.6-503c155.6-204.6,292-373,300.5-373c8.5,0,157.7,185.4,332.5,413.5c334.6,439.1,496.6,598.9,767.3,760.9c1052.9,626.6,2408.5,223.8,2949.9-876c136.4-275,191.8-492.4,206.8-799.3c19.2-396.4-57.6-701.2-275-1078.5C8746.2,390.4,8650.3,286,6787.5-1579L5001.4-3369.4L3225.9-1589.6C2168.7-530.3,1390.7,264.7,1307.6,375.5C1022,744.3,836.6,1113,766.2,1454c-44.8,219.5-27.7,609.6,36.2,846.2c183.3,682.1,716.2,1240.5,1385.4,1455.8C2531,3866.8,2906.2,3886,3253.6,3813.5z" />
+			</g>
+		</g>
+	</svg>
+)
